requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.1
boto3==1.37.23
botocore==1.37.23
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.1
google-auth==2.38.0
griffe==1.7.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.0
idna==3.10
importlib_metadata==8.6.1
jiter==0.9.0
jmespath==1.0.1
logfire-api==3.12.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
openai==1.69.0
opentelemetry-api==1.31.1
packaging==24.2
prompt_toolkit==3.0.50
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-ai==0.0.46
pydantic-ai-slim==0.0.46
pydantic-graph==0.0.46
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
moviepy>=1.0.3
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.24.0
ffmpeg-python>=0.2.0
tqdm>=4.65.0
setuptools>=65.5.1  # Required by moviepy
decorator>=4.0.2  # Required by moviepy
imageio>=2.5.0  # Required by moviepy
imageio-ffmpeg>=0.4.4  # Required by moviepy
easyocr>=1.7.0
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
torchvision>=0.15.0
fastapi
supabase==2.11.0
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
aiofiles
PyPDF2
docx2txt
python-multipart
python-docx
pydub
